pyproject.toml
[tool.poetry]
name = "coin_bot"
version = "0.1.0"
description = ""
authors = ["Aleksandr Volkov <jetvolkov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.16.1"
environs = "^9.3.0"
python-telegram-bot = "^13.1"
Babel = "^2.9.0"
peewee = "^3.14.0"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
